requirements.txt
# requirements.txt
streamlit==1.29.0
google-generativeai==0.8.3  # Gemini API SDK
google-ai-generativelanguage>=0.6.18,<0.7.0  # Resolve dependency conflict
textblob==0.17.1
nltk==3.8.1  # Required by TextBlob
